Move different Bloom impls into their own submodules
